WL#16659 MySQL Shell option tracker registration
This patch introduces the logic to have the shell core functionality
reported in the option tracker including:

- Shell usage (per user session)
- Dump utility usage (per dump utility execution)
- Dump Load usage
- Upgrade Checker usage
- Copy usage (per copy utility execution)

Supports defining the execution context and thus, reporting the above
options in the correct category when used from VSCode extension.

Additionally, implements a set_report_option in the session object
to enable reporting of custom tracking options.

Change-Id: Ib60417bbf8596abf99afac904c7d6ee2e10232c2