BUG#38560511 MysqlShell failed to replace default collation from 'utf8mb4_uca1400_ai_ci' to 'utf8mb4_0900_ai_ci'
Dumping a MariaDB table that had an explicitly specified collation and
contained a string column without a collation clause failed with the
following error:

  Failed to replace default collation of table `sakila`.`ptable` from
  'utf8mb4_uca1400_ai_ci' to 'utf8mb4_0900_ai_ci'.

If a string column does not have a collation specified, it uses the
table's collation. In such cases, if the table's collation is the
default collation for its character set, it will not appear in the
output of the SHOW CREATE TABLE statement. When the dumper attempted to
replace unsupported collations with their closest counterparts, it
expected to find the collation specified twice: once for the column and
once for the table. Since only one collation clause was present, it
reported an error when the second occurrence was missing.

As a fix, the presence of collation in the given statement is no longer
verified.

Change-Id: I4ed2d4bf0be0bacd28a1be4dcc8ca6653fba7960