Fix check which verifies if metadata/DDL files were written
Ensure that the success condition of check which verifies if all
metadata and DDL files have been written does not depend on variables
which can be modified while this check is being executed.

The `m_all_table_metadata_tasks_scheduled` has been removed, as it was
possible that metadata and DDL tasks were finished before it was set to
true, preventing i.e. DDL errors from aborting the dump.

Change-Id: Ia8bc365caa14bf420c3e920beb1cc99a3c08c17f