BUG#38629024 Improve performance of gathering list of views
Gathering the list of views from an instance that does not have the
information_schema.view_table_usage table was slow when the number of
views was high.

The issue was caused by the slow execution of a section of code that
extracted table references from view definitions.

Performance has been improved by changing the prediction mode used by
the parser to a faster one, which better handles short, well formed
queries. The other prediction mode is used as a fall-back.

Additionally, added caching of some of the objects used during the
extraction process.

Change-Id: I07707074e90cd8908662f81a568fe439c09155b3