BUG#38555376 The 'syntax' check fails with a 'Unknown database' error
This patch introduces the capacity of some SQL checks to automatically
exclude database objects with errors from further checks.

With this change, the orphaned objects check is moved to the beginning
of the list, any orphaned object will be automatically excluded from
the rest of the checks, by adding an exclussion on the filter.

Change-Id: I849cc33ce3dd1f5c10b3279f30a061ff515dbcd3