BUG#38650807 Verify if it's possible to change table's engine to InnoDB
When running a dump with the 'ocimds' option enabled, if table is not
using the InnoDB engine, an additional check is executed to verify if
it's possible to change table's engine to InnoDB. If this check fails,
an error is reported, which requires a manual fix.

The CREATE TEMPORARY TABLES privilege is now required when running the
dump.

Additionally, handling of the 'force_innodb' compatibility option has
been improved when dumping without 'ocimds':true option:

 - DATA|INDEX DIRECTORY clause is automatically commented out
 - an error is reported if table has more columns than InnoDB engine
   supporst

Change-Id: Iff19dbeda4beae1ee4771c6031d5ecb4eb401791