Fix failing test
The `Utils_private_key_test.load_save_and_read` test was sometimes
failing when the wrong password check was throwing an unexpected
exception.

Reading an encrypted private key may fail either when decryption fails,
or when decoding fails. In certain situations decryption may succeed
even though a wrong password was provided. In such case a decoding
error is reported, and check for that error was missing.

Change-Id: I5d287022ac8b26c1d13fb8a2f269fd3dd2ae80ff