BUG#38628445 Improve performance of UC queries
The schemaInconsistency check was updated by including the schema and routine
filters in the query, resulting in a moderate improvement.

The orphanedObjects check was improved by replacing the derived query with
a left join resulting in a significant improvement.

The query for the indexTooLarge was updated to use the i_s.tables table
instead of i_s.innodb_sys_tables to remove the concat operation in place
which resulted in significant improvement.

The 2 heavy queries for deprecatedTemporalDelimiter have been merged into a
single one to avoid the overhead.

Change-Id: I0287aa53e05af8bebab8535073e7f36ddfee5fd4