BUG#38748787 Output of spatialIndex check contains unescaped placeholder
The problem is caused by 2 reasons: a bad formatting was used in the messages
file, and the value for the target_version token was not defined for the check.

The problem is fixed by introducing a base_tokens() function that would return
tokens that are applicable for the resolution of descriptions in all checks,
for the moment, only the Sql_upgrade_check returns the token for the target_version
which is then used on the resolution of the spatialIndex check.

Change-Id: I04913263a2598b04cda9fec3f9304c118019d002